# docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  theme.ThemeConfigRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      theme_name:
        type: string
    required:
    - config
    - theme_name
    type: object
...
  theme.ThemeInfo:
    properties:
      author:
        type: string
      createdAt:
        type: string
      demoUrl:
        type: string
      description:
        type: string
      downloadCount:
        type: integer
      downloadUrl:
        type: string
      id:
        type: integer
      install_time:
        description: 安装时间
        type: string
      installed_version:
        description: 已安装版本
        type: string
      instructionUrl:
        type: string
      is_current:
        description: 本地状态字段
        type: boolean
      is_installed:
        description: 是否已安装（对于已安装主题列表始终为true）
        type: boolean
      isActive:
        type: boolean
      isOfficial:
        type: boolean
      name:
        type: string
      previewUrl:
        type: string
      price:
        type: integer
      rating:
        type: number
      repoUrl:
        type: string
      tags:
        items:
          type: string
        type: array
      themeType:
        type: string
      updatedAt:
        type: string
      user_config:
        additionalProperties: true
        description: 用户配置
        type: object
      version:
        type: string
    type: object
  theme.ThemeMarketListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/theme.MarketTheme'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /public/theme/config:
    get:
      description: 获取当前激活主题的配置值（供前端主题使用，只返回配置值）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取当前主题配置（公开）
      tags:
      - 主题配置
  /public/theme/market:
    get:
      description: 获取主题商城中的所有可用主题（PRO 版本会返回包含完整 downloadUrl 的 PRO 主题）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.ThemeMarketListResponse'
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取主题商城列表
      tags:
      - 主题商城
...
  /theme/config:
    get:
      description: 获取用户对指定主题的配置值
      parameters:
      - description: 主题名称
        in: query
        name: theme_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户主题配置
      tags:
      - 主题配置
    post:
      consumes:
      - application/json
      description: 保存用户对指定主题的配置值
      parameters:
      - description: 主题配置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/theme.ThemeConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 保存成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 保存失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 保存用户主题配置
      tags:
      - 主题配置
  /theme/current:
    get:
      description: 获取用户当前使用的主题信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.ThemeInfo'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前主题
      tags:
      - 主题管理
  /theme/current-config:
    get:
      description: 获取当前激活主题的配置定义和值（供前端主题使用的公开接口）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.ThemeConfigResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前主题配置
      tags:
      - 主题配置
  /theme/fix-status:
    post:
      description: 修复用户主题的当前状态数据一致性，解决多个主题同时标记为当前使用的问题
      produces:
      - application/json
      responses:
        "200":
          description: 修复成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 修复失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修复主题状态
      tags:
      - 主题管理
  /theme/install:
    post:
      consumes:
      - application/json
      description: 从指定URL下载并安装主题（主题名必须以theme-开头）
      parameters:
      - description: 主题安装请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_anzhiyu-c_anheyu-app_pkg_service_theme.ThemeInstallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 安装成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 安装失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 安装主题
      tags:
      - 主题管理
  /theme/installed:
    get:
      description: 获取用户已安装的所有主题
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/theme.ThemeInfo'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取已安装主题列表
      tags:
      - 主题管理
  /theme/official:
    post:
      description: 切换到官方内嵌主题
      produces:
      - application/json
      responses:
        "200":
          description: 切换成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 切换失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 切换到官方主题
      tags:
      - 主题管理
  /theme/settings:
    get:
      description: 获取指定主题的配置字段定义（用于后台生成配置表单）
      parameters:
      - description: 主题名称
        in: query
        name: theme_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/theme.ThemeSettingGroup'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取主题配置定义
      tags:
      - 主题配置
  /theme/static-mode:
    get:
      description: 检查当前是否处于静态主题模式（是否存在static目录）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.StaticModeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 检查静态模式
      tags:
      - 主题管理
  /theme/switch:
    post:
      consumes:
      - application/json
      description: 切换到指定的已安装主题或官方主题
      parameters:
      - description: 切换主题请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/theme.SwitchThemeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 切换成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 切换失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 切换主题
      tags:
      - 主题管理
  /theme/uninstall:
    post:
      consumes:
      - application/json
      description: 卸载指定的主题（不能卸载当前使用的主题）
      parameters:
      - description: 卸载主题请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/theme.UninstallThemeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 卸载成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 卸载失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 卸载主题
      tags:
      - 主题管理
  /theme/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传主题压缩包文件（ZIP格式，最大50MB），系统会自动解析theme.json并安装主题
      parameters:
      - description: 主题压缩包文件
        in: formData
        name: file
        required: true
        type: file
      - description: 是否强制更新
        in: formData
        name: force_update
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.ThemeUploadResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 上传失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 上传主题压缩包
      tags:
      - 主题管理
  /theme/validate:
    post:
      consumes:
      - multipart/form-data
      description: 验证主题压缩包的格式和内容是否符合规范
      parameters:
      - description: 主题压缩包文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 验证成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 验证失败
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 验证主题压缩包
      tags:
      - 主题管理
...
swagger: "2.0"